[UPDATE] production version agent
